requirements.txt
numpy
pandas
pyaml
seaborn
scikit-learn
matplotlib
mysql-connector-python